+++ table/block.h
@@ -18,7 +18,6 @@
 #include "db/pinned_iterators_manager.h"
 #include "rocksdb/iterator.h"
 #include "rocksdb/options.h"
-#include "table/block_hash_index.h"
 #include "table/block_prefix_index.h"
 #include "table/internal_iterator.h"
 
@@ -29,7 +28,6 @@ namespace rocksdb {
 struct BlockContents;
 class Comparator;
 class BlockIter;
-class BlockHashIndex;
 class BlockPrefixIndex;
 
 class Block {
@@ -71,7 +69,6 @@ class Block {
   InternalIterator* NewIterator(const Comparator* comparator,
                                 BlockIter* iter = nullptr,
                                 bool total_order_seek = true);
-  void SetBlockHashIndex(BlockHashIndex* hash_index);
   void SetBlockPrefixIndex(BlockPrefixIndex* prefix_index);
 
   // Report an approximation of how much memory has been used.
@@ -82,7 +79,6 @@ class Block {
   const char* data_;            // contents_.data.data()
   size_t size_;                 // contents_.data.size()
   uint32_t restart_offset_;     // Offset in data_ of restart array
-  std::unique_ptr<BlockHashIndex> hash_index_;
   std::unique_ptr<BlockPrefixIndex> prefix_index_;
 
   // No copying allowed
@@ -100,20 +96,17 @@ class BlockIter : public InternalIterator {
         current_(0),
         restart_index_(0),
         status_(Status::OK()),
-        hash_index_(nullptr),
         prefix_index_(nullptr) {}
 
   BlockIter(const Comparator* comparator, const char* data, uint32_t restarts,
-       uint32_t num_restarts, BlockHashIndex* hash_index,
-       BlockPrefixIndex* prefix_index)
+            uint32_t num_restarts, BlockPrefixIndex* prefix_index)
       : BlockIter() {
-    Initialize(comparator, data, restarts, num_restarts,
-        hash_index, prefix_index);
+    Initialize(comparator, data, restarts, num_restarts, prefix_index);
   }
 
   void Initialize(const Comparator* comparator, const char* data,
-      uint32_t restarts, uint32_t num_restarts, BlockHashIndex* hash_index,
-      BlockPrefixIndex* prefix_index) {
+                  uint32_t restarts, uint32_t num_restarts,
+                  BlockPrefixIndex* prefix_index) {
     assert(data_ == nullptr);           // Ensure it is called only once
     assert(num_restarts > 0);           // Ensure the param is valid
 
@@ -123,7 +116,6 @@ class BlockIter : public InternalIterator {
     num_restarts_ = num_restarts;
     current_ = restarts_;
     restart_index_ = num_restarts_;
-    hash_index_ = hash_index;
     prefix_index_ = prefix_index;
   }
 
@@ -179,7 +171,6 @@ class BlockIter : public InternalIterator {
   IterKey key_;
   Slice value_;
   Status status_;
-  BlockHashIndex* hash_index_;
   BlockPrefixIndex* prefix_index_;
 
   inline int Compare(const Slice& a, const Slice& b) const {
@@ -219,8 +210,6 @@ class BlockIter : public InternalIterator {
   bool BinaryBlockIndexSeek(const Slice& target, uint32_t* block_ids,
                             uint32_t left, uint32_t right,
                             uint32_t* index);
-
-  bool HashSeek(const Slice& target, uint32_t* index);
 
   bool PrefixSeek(const Slice& target, uint32_t* index);